// integrations/instagram/src/schema.ts
export type InstagramAuthValue = Oauth2AuthValue & {
  metadata: {
    igId: string
    igName: string
    pageId: string
    version: string
  }
}
// ...
export const iceBreakerSchema = z.object({
  locale: z.string(),
  call_to_actions: z.array(
    z.object({
      question: z.string(),
      payload: z.string(),
    }),
  ),
})
export type IceBreaker = z.infer<typeof iceBreakerSchema>

export const persistentMenuSchema = z.object({
  locale: z.string(),
  call_to_actions: z.array(instagramButtonSchema),
})
export type PersistentMenuSchema = z.infer<typeof persistentMenuSchema>

export const instagramProfileRequest = z.object({
  ice_breakers: z.array(iceBreakerSchema),
  persistent_menu: z.array(persistentMenuSchema),
})
export type InstagramProfileRequest = z.infer<typeof instagramProfileRequest>